+++ cli/src/test/download.spec.ts
@@ -1,7 +1,7 @@
 import anyTest, { TestInterface } from 'ava';
 import * as td from 'testdouble';
 import { Context, addFake } from './fake_lycan';
-import { success } from 'specshell';
+import { success, Output, Exit, Signal } from 'specshell';
 import * as tar from 'tar';
 
 const test = anyTest as TestInterface<Context>;
@@ -15,6 +15,15 @@ const defaultApp = {
   name: 'fake-name',
   environments: [],
 };
+
+const stabilize = (result: Output<string> & (Exit | Signal)) => ({
+  ...result,
+  out: result.out
+    .replace(/up to date in [0-9][.0-9]*s/g, 'up to date')
+    .replace(/localhost:[0-9]*/, 'localhost'),
+  err: result.err
+    .replace(/localhost:[0-9]*/, 'localhost'),
+});
 
 const anything = td.matchers.anything();
 
@@ -84,6 +93,8 @@ test('bad url', async (t) => {
       githubUrl: 'someUrl',
       downloadUrl: `${t.context.lycanUrl}/bad-url/foo`,
       downloadDir: 'someDir',
+      targetDir: 'someDir',
+      zipUrl: 'xxx',
     },
   });
   (t.context.lycanServer as any).router.koaRouter.get('/bad-url/foo', (ctx: any) => {
@@ -91,11 +102,7 @@ test('bad url', async (t) => {
   });
 
   const result = await t.context.shell.run(`${t.context.run} download app-with-bad-url`, 'utf-8');
-  const stableResult = {
-    ...result,
-    err: result.err.replace(/localhost:[0-9]*/, 'localhost'),
-  };
-  t.snapshot(stableResult);
+  t.snapshot(stabilize(result));
 });
 
 test('bad tgz', async (t) => {
@@ -106,6 +113,8 @@ test('bad tgz', async (t) => {
       githubUrl: 'someUrl',
       downloadUrl: `${t.context.lycanUrl}/bad-tgz/archive/master.tar.gz`,
       downloadDir: 'bad-tgz',
+      targetDir: 'bad-tgz',
+      zipUrl: 'xxx',
     },
   });
   const tarStream = tar.create({ gzip: true, cwd: 'src/test/apps' }, ['bad-tgz']);
@@ -127,6 +136,8 @@ test('verbose bad tgz', async (t) => {
       githubUrl: 'someUrl',
       downloadUrl: `${t.context.lycanUrl}/bad-tgz/archive/master.tar.gz`,
       downloadDir: 'bad-tgz',
+      targetDir: 'bad-tgz',
+      zipUrl: 'xxx',
     },
   });
   const tarStream = tar.create({ gzip: true, cwd: 'src/test/apps' }, ['bad-tgz']);
@@ -147,7 +158,9 @@ test('bad request', async (t) => {
     source: {
       githubUrl: 'someUrl',
       downloadUrl: `${t.context.lycanUrl}/bad-req/archive/master.tar.gz`,
-      downloadDir: 'bad-req',
+      downloadDir: 'bad-req-master',
+      targetDir: 'bad-req',
+      zipUrl: 'xxx',
     },
   });
   (t.context.lycanServer as any).router.koaRouter.get('/bad-req/archive/master.tar.gz', async (ctx: any) => {
@@ -155,11 +168,7 @@ test('bad request', async (t) => {
   });
 
   const result = await t.context.shell.run(`${t.context.run} download app-with-bad-req`, 'utf-8');
-  const stableResult = {
-    ...result,
-    err: result.err.replace(/localhost:[0-9]*/, 'localhost'),
-  };
-  t.snapshot(stableResult);
+  t.snapshot(stabilize(result));
 });
 
 test('verbose bad request', async (t) => {
@@ -170,6 +179,8 @@ test('verbose bad request', async (t) => {
       githubUrl: 'someUrl',
       downloadUrl: `${t.context.lycanUrl}/bad-req/archive/master.tar.gz`,
       downloadDir: 'bad-req',
+      targetDir: 'bad-req',
+      zipUrl: 'xxx',
     },
   });
   (t.context.lycanServer as any).router.koaRouter.get('/bad-req/archive/master.tar.gz', async (ctx: any) => {
@@ -177,12 +188,7 @@ test('verbose bad request', async (t) => {
   });
 
   const result = await t.context.shell.run(`${t.context.run} download -v app-with-bad-req`, 'utf-8');
-  const stableResult = {
-    ...result,
-    out: result.out.replace(/localhost:[0-9]*/, 'localhost'),
-    err: result.err.replace(/localhost:[0-9]*/, 'localhost'),
-  };
-  t.snapshot(stableResult);
+  t.snapshot(stabilize(result));
 });
 
 test('tgz with wrong dir', async (t) => {
@@ -193,6 +199,8 @@ test('tgz with wrong dir', async (t) => {
       githubUrl: 'foo',
       downloadUrl: `${t.context.lycanUrl}/bad-dir/archive/master.tar.gz`,
       downloadDir: 'not-good',
+      targetDir: 'still-not-good',
+      zipUrl: 'xxx',
     },
   });
   const tarStream = tar.create({ gzip: true, cwd: 'src/test/apps' }, ['good-app-master']);
@@ -202,11 +210,73 @@ test('tgz with wrong dir', async (t) => {
   });
 
   const result = await t.context.shell.run(`${t.context.run} download bad-dir`, 'utf-8');
-  const stableResult = {
-    ...result,
-    out: result.out.replace(/up to date in [0-9][.0-9]*s/g, 'up to date'),
-  };
-  t.snapshot(stableResult);
+  t.snapshot(stabilize(result));
+});
+
+test('good tgz existing target file', async (t) => {
+  const targetDir = 'fubar';
+  td.when(t.context.lycanFake.getApp(anything, 'exist-app')).thenResolve({
+    ...defaultApp,
+    id: 'exist-app',
+    source: {
+      githubUrl: 'foo',
+      downloadUrl: `${t.context.lycanUrl}/exist-app/archive/master.tar.gz`,
+      downloadDir: 'good-app-master',
+      targetDir,
+      zipUrl: 'xxx',
+    },
+  });
+  const tarStream = tar.create({ gzip: true, cwd: 'src/test/apps' }, ['good-app-master']);
+  const tgzBuffer = await drainToBuffer(tarStream);
+  (t.context.lycanServer as any).router.koaRouter.get('/exist-app/archive/master.tar.gz', (ctx: any) => {
+    ctx.body = tgzBuffer;
+  });
+  const result = await t.context.shell.run(`touch ${targetDir} && ${t.context.run} download exist-app`, 'utf-8');
+  t.snapshot(stabilize(result));
+});
+
+test('good tgz existing non empty target dir', async (t) => {
+  const targetDir = 'fubar';
+  td.when(t.context.lycanFake.getApp(anything, 'exist-app')).thenResolve({
+    ...defaultApp,
+    id: 'exist-app',
+    source: {
+      githubUrl: 'foo',
+      downloadUrl: `${t.context.lycanUrl}/exist-app/archive/master.tar.gz`,
+      downloadDir: 'good-app-master',
+      targetDir,
+      zipUrl: 'xxx',
+    },
+  });
+  const tarStream = tar.create({ gzip: true, cwd: 'src/test/apps' }, ['good-app-master']);
+  const tgzBuffer = await drainToBuffer(tarStream);
+  (t.context.lycanServer as any).router.koaRouter.get('/exist-app/archive/master.tar.gz', (ctx: any) => {
+    ctx.body = tgzBuffer;
+  });
+  const result = await t.context.shell.run(`mkdir -p ${targetDir}/x && ${t.context.run} download exist-app`, 'utf-8');
+  t.snapshot(stabilize(result));
+});
+
+test('good tgz existing empty target dir', async (t) => {
+  const targetDir = 'fubar';
+  td.when(t.context.lycanFake.getApp(anything, 'exist-app')).thenResolve({
+    ...defaultApp,
+    id: 'exist-app',
+    source: {
+      githubUrl: 'foo',
+      downloadUrl: `${t.context.lycanUrl}/exist-app/archive/master.tar.gz`,
+      downloadDir: 'good-app-master',
+      targetDir,
+      zipUrl: 'xxx',
+    },
+  });
+  const tarStream = tar.create({ gzip: true, cwd: 'src/test/apps' }, ['good-app-master']);
+  const tgzBuffer = await drainToBuffer(tarStream);
+  (t.context.lycanServer as any).router.koaRouter.get('/exist-app/archive/master.tar.gz', (ctx: any) => {
+    ctx.body = tgzBuffer;
+  });
+  const result = await t.context.shell.run(`mkdir ${targetDir} && ${t.context.run} download exist-app`, 'utf-8');
+  t.snapshot(stabilize(result));
 });
 
 test('good tgz', async (t) => {
@@ -217,6 +287,8 @@ test('good tgz', async (t) => {
       githubUrl: 'foo',
       downloadUrl: `${t.context.lycanUrl}/good-app/archive/master.tar.gz`,
       downloadDir: 'good-app-master',
+      targetDir: 'good-app',
+      zipUrl: 'xxx',
     },
   });
   const tarStream = tar.create({ gzip: true, cwd: 'src/test/apps' }, ['good-app-master']);
@@ -226,9 +298,5 @@ test('good tgz', async (t) => {
   });
 
   const result = await t.context.shell.run(`${t.context.run} download good-app`, 'utf-8');
-  const stableResult = {
-    ...result,
-    out: result.out.replace(/up to date in [0-9][.0-9]*s/g, 'up to date'),
-  };
-  t.snapshot(stableResult);
+  t.snapshot(stabilize(result));
 });