+++ leveldb-server/src/test/client.test.ts
@@ -6,18 +6,23 @@ import * as path from 'path';
 import { Handler } from '../db';
 import { setUpTests } from '@reshuffle/db-testsuite/dist/client';
 
-setUpTests<{ dbDir: string }>({
+setUpTests<{ dbDir: string, errorsOk?: boolean, errors: Error[] }>({
   async setUp() {
+    const errors: Error[] = [];
     const dbDir = await promisify(mkdtemp)(path.join(tmpdir(), 'test-state-'), 'utf8');
     return {
-      handler: new Handler(`${dbDir}/root.db`),
+      handler: new Handler(`${dbDir}/root.db`, (err) => err && errors.push(err)),
       supportsPolling: true,
       context: {
         dbDir,
+        errors,
       },
     };
   },
   async tearDown(ctx) {
+    if (!ctx.errorsOk && ctx.errors.length > 0) {
+      throw new Error(`Unexamined errors: ${ctx.errors}`);
+    }
     await rmrf(ctx.dbDir);
   },
 });